encoder_service/requirements.txt
absl-py==2.4.0
accelerate==1.12.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.0
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
bleach==6.2.0
calflops==0.3.2
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cvzone==1.6.1
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
executing==2.2.1
fastapi==0.129.0
faster-coco-eval==1.7.1
fastjsonschema==2.21.2
filelock==3.20.0
filetype==1.2.0
fonttools==4.60.1
fqdn==1.5.1
fsspec==2025.9.0
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.0
idna==3.7
Jinja2==3.1.6
json5==0.12.1
jsonpointer==3.0.0
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.15
kiwisolver==1.4.9
labelImg==1.8.6
lark==1.3.0
loguru==0.7.3
lxml==6.0.2
Markdown==3.10.1
MarkupSafe==3.0.3
matplotlib==3.10.7
mistune==3.1.4
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.2.6
nvidia-ml-py3==7.352.0
opencv-python==4.12.0.88
opencv-python-headless==4.10.0.84
packaging==25.0
pandocfilters==1.5.1
parso==0.8.5
pi_heif==1.1.1
pillow==11.3.0
pillow-avif-plugin==1.5.2
platformdirs==4.5.0
polars==1.34.0
polars-runtime-32==1.34.0
prometheus_client==0.23.1
protobuf==6.33.5
psutil==7.1.0
pure_eval==0.2.3
pycocotools==2.0.11
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==4.0.0
python-multipart==0.0.22
pywin32==311
pywinpty==3.0.2
PyYAML==6.0.3
pyzmq==27.1.0
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
roboflow==1.2.10
rpds-py==0.27.1
safetensors==0.7.0
scipy==1.16.2
Send2Trash==1.8.3
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
starlette==0.52.1
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
terminado==0.18.1
tinycss2==1.4.0
tokenizers==0.22.2
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==5.1.0
typer-slim==0.21.1
types-python-dateutil==2.9.0.20251008
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.3.232
ultralytics-thop==2.0.18
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.40.0
wcwidth==0.2.14
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.5
widgetsnbextension==4.0.14
win32_setctime==1.2.0
